Move contact to pages
